Separate app and server so it won't keep jest from exiting
